Complement to sticky patch: Ignore sticky if fullscreen
With the original patch, when a sticky application is fullscreen,
switching tags has no effect, as the application is present (and
fullscreen) in all tags.

While this behaviour technically makes sense, I don't find it useful.
Why would I want to stick a fullscreen window?

So this commit ignores the sticky flag whenever the window is
fullscreen. This way, the window is only present (and fullscreen) on its
own tag, and we get the same behaviour as with a non-sticky window, such
as changing tags normally.